Final testing and selecting a tabbing system.
