+++ src/renderer/src/model/splits.ts
@@ -1,0 +1,26 @@
+import { Tab } from "./tabs";
+import { ContentDirection } from "./view";
+
+
+export type SplitSide = "left" | "right";
+
+export interface SplitTreeItem {
+  isFork: boolean;
+  side: SplitSide | null;
+}
+
+export type SplitTreeNode = SplitTreeFork | SplitTreeValue;
+
+export interface SplitTreeFork extends SplitTreeItem {
+  direction: ContentDirection;
+  left: SplitTreeNode;
+  right?: SplitTreeNode | null;
+}
+
+export interface SplitTreeValue extends SplitTreeItem {
+  value: Tab[];
+}
+
+export type SplitTree = {
+  root: SplitTreeNode;
+};
