+++ src/renderer/src/model/tabs.ts
@@ -1,0 +1,8 @@
+import { ReactNode } from "react";
+
+export type Tab = {
+  id: string;
+  workspace: string;
+  caption: string;
+  content: ReactNode;
+};
